began moving gfx demo into own isolated game
