+++ src/games/mod.rs
@@ -1,0 +1,1 @@
+pub mod light_gfx_demo;
